requirements.txt
tqdm==4.56.0
git+https://github.com/openai/CLIP.git
requests==2.25.1
streamlit==0.74.1
torch==1.7.1
numpy==1.20.0rc2
Pillow==8.1.0